seprate newTaskModel
